+++ keydive/__init__.py
@@ -4,4 +4,4 @@ from .cdm import Cdm
 from .vendor import Vendor
 from .keybox import Keybox
 
-__version__ = '2.1.5'
+__version__ = "2.2.0"