+++ keydive/constants.py
@@ -1,143 +1,145 @@
-import re
 from pathlib import Path
 
 from keydive.vendor import Vendor
 
-# https://developer.android.com/ndk/guides/cpp-support
+# https://cs.android.com/android/platform/superproject/+/android14-qpr3-release:bionic/libc/libc.map.txt
 NATIVE_C_API = {
     # BUILT-IN
-    'main',
+    "main",
     # STDIO
-    'fclose', 'fflush', 'fgetc', 'fgetpos', 'fgets', 'fopen', 'fprintf', 'fputc', 'fputs', 'fread', 'freopen',
-    'fscanf', 'fseek', 'fsetpos', 'ftell', 'fwrite', 'getc', 'getchar', 'gets', 'perror', 'printf', 'putc',
-    'putchar', 'puts', 'remove', 'rename', 'rewind', 'scanf', 'setbuf', 'setvbuf', 'sprintf', 'sscanf', 'tmpfile',
-    'tmpnam', 'ungetc', 'vfprintf', 'vprintf', 'vsprintf', 'fileno', 'feof', 'ferror', 'snprintf',
+    "fclose", "fflush", "fgetc", "fgetpos", "fgets", "fopen", "fprintf", "fputc", "fputs", "fread", "freopen",
+    "fscanf", "fseek", "fsetpos", "ftell", "fwrite", "getc", "getchar", "gets", "perror", "printf", "putc",
+    "putchar", "puts", "remove", "rename", "rewind", "scanf", "setbuf", "setvbuf", "sprintf", "sscanf", "tmpfile",
+    "tmpnam", "ungetc", "vfprintf", "vprintf", "vsprintf", "fileno", "feof", "ferror", "snprintf",
     # STDLIB
-    'abort', 'abs', 'atexit', 'atof', 'atoi', 'atol', 'bsearch', 'calloc', 'div', 'exit', 'free', 'getenv', 'labs',
-    'ldiv', 'malloc', 'mblen', 'mbstowcs', 'mbtowc', 'qsort', 'rand', 'realloc', 'srand', 'strtod', 'strtol',
-    'strtoul', 'system', 'wcstombs', 'wctomb',
+    "abort", "abs", "atexit", "atof", "atoi", "atol", "bsearch", "calloc", "div", "exit", "free", "getenv", "labs",
+    "ldiv", "malloc", "mblen", "mbstowcs", "mbtowc", "qsort", "rand", "realloc", "srand", "strtod", "strtol",
+    "strtoul", "system", "wcstombs", "wctomb",
     # STRING
-    'memchr', 'memcmp', 'memcpy', 'memmove', 'memset', 'strcat', 'strchr', 'strcmp', 'strcoll', 'strcpy', 'strcspn',
-    'strerror', 'strlen', 'strncat', 'strncmp', 'strncpy', 'strpbrk', 'strrchr', 'strspn', 'strstr', 'strtok',
-    'strxfrm', 'strncasecmp',
+    "memchr", "memcmp", "memcpy", "memmove", "memset", "strcat", "strchr", "strcmp", "strcoll", "strcpy", "strcspn",
+    "strerror", "strlen", "strncat", "strncmp", "strncpy", "strpbrk", "strrchr", "strspn", "strstr", "strtok",
+    "strxfrm", "strncasecmp",
     # MATH
-    'acos', 'asin', 'atan', 'atan2', 'cos', 'cosh', 'exp', 'fabs', 'floor', 'fmod', 'frexp', 'ldexp', 'log',
-    'log10', 'modf', 'pow', 'sin', 'sinh', 'sqrt', 'tan', 'tanh',
+    "acos", "asin", "atan", "atan2", "cos", "cosh", "exp", "fabs", "floor", "fmod", "frexp", "ldexp", "log",
+    "log10", "modf", "pow", "sin", "sinh", "sqrt", "tan", "tanh",
     # CTYPE
-    'isalnum', 'isalpha', 'iscntrl', 'isdigit', 'isgraph', 'islower', 'isprint', 'ispunct', 'isspace', 'isupper',
-    'isxdigit', 'tolower', 'toupper',
+    "isalnum", "isalpha", "iscntrl", "isdigit", "isgraph", "islower", "isprint", "ispunct", "isspace", "isupper",
+    "isxdigit", "tolower", "toupper",
     # TIME
-    'asctime', 'clock', 'ctime', 'difftime', 'gmtime', 'localtime', 'mktime', 'strftime', 'time',
+    "asctime", "clock", "ctime", "difftime", "gmtime", "localtime", "mktime", "strftime", "time",
     # UNISTD
-    'access', 'alarm', 'chdir', 'chown', 'close', 'dup', 'dup2', 'execle', 'execv', 'execve', 'execvp', 'fork',
-    'fpathconf', 'getcwd', 'getegid', 'geteuid', 'getgid', 'getgroups', 'getlogin', 'getopt', 'getpgid', 'getpgrp',
-    'getpid', 'getppid', 'getuid', 'isatty', 'lseek', 'pathconf', 'pause', 'pipe', 'read', 'rmdir', 'setgid',
-    'setpgid', 'setsid', 'setuid', 'sleep', 'sysconf', 'tcgetpgrp', 'tcsetpgrp', 'ttyname', 'ttyname_r', 'write',
-    'fsync', 'unlink', 'syscall', 'getpagesize',
+    "access", "alarm", "chdir", "chown", "close", "dup", "dup2", "execle", "execv", "execve", "execvp", "fork",
+    "fpathconf", "getcwd", "getegid", "geteuid", "getgid", "getgroups", "getlogin", "getopt", "getpgid", "getpgrp",
+    "getpid", "getppid", "getuid", "isatty", "lseek", "pathconf", "pause", "pipe", "read", "rmdir", "setgid",
+    "setpgid", "setsid", "setuid", "sleep", "sysconf", "tcgetpgrp", "tcsetpgrp", "ttyname", "ttyname_r", "write",
+    "fsync", "unlink", "syscall", "getpagesize",
     # FCNTL
-    'creat', 'fcntl', 'open',
+    "creat", "fcntl", "open",
     # SYS_TYPE
-    'fd_set', 'FD_CLR', 'FD_ISSET', 'FD_SET', 'FD_ZERO',
+    "fd_set", "FD_CLR", "FD_ISSET", "FD_SET", "FD_ZERO",
     # SYS_STAT
-    'chmod', 'fchmod', 'fstat', 'mkdir', 'mkfifo', 'stat', 'umask',
+    "chmod", "fchmod", "fstat", "mkdir", "mkfifo", "stat", "umask",
     # SYS_TIME
-    'gettimeofday', 'select', 'settimeofday',
+    "gettimeofday", "select", "settimeofday",
     # SIGNAL
-    'signal', 'raise', 'kill', 'sigaction', 'sigaddset', 'sigdelset', 'sigemptyset', 'sigfillset', 'sigismember',
-    'sigpending', 'sigprocmask', 'sigsuspend', 'alarm', 'pause',
+    "signal", "raise", "kill", "sigaction", "sigaddset", "sigdelset", "sigemptyset", "sigfillset", "sigismember",
+    "sigpending", "sigprocmask", "sigsuspend", "alarm", "pause",
     # SETJMP
-    'longjmp', 'setjmp',
+    "longjmp", "setjmp",
     # ERRNO
-    'errno', 'strerror', 'perror',
+    "errno", "strerror", "perror",
     # ASSERT
-    'assert',
+    "assert",
     # LOCAL
-    'localeconv', 'setlocale',
+    "localeconv", "setlocale",
     # WCHAR
-    'btowc', 'fgetwc', 'fgetws', 'fputwc', 'fputws', 'fwide', 'fwprintf', 'fwscanf', 'getwc', 'getwchar', 'mbrlen',
-    'mbrtowc', 'mbsinit', 'mbsrtowcs', 'putwc', 'putwchar', 'swprintf', 'swscanf', 'ungetwc', 'vfwprintf',
-    'vfwscanf', 'vwprintf', 'vwscanf', 'wcrtomb', 'wcscat', 'wcschr', 'wcscmp', 'wcscoll', 'wcscpy', 'wcscspn',
-    'wcsftime', 'wcslen', 'wcsncat', 'wcsncmp', 'wcsncpy', 'wcspbrk', 'wcsrchr', 'wcsrtombs', 'wcsspn', 'wcsstr',
-    'wcstod', 'wcstok', 'wcstol', 'wcstombs', 'wcstoul', 'wcsxfrm', 'wctob', 'wmemchr', 'wmemcmp', 'wmemcpy',
-    'wmemmove', 'wmemset', 'wprintf', 'wscanf',
+    "btowc", "fgetwc", "fgetws", "fputwc", "fputws", "fwide", "fwprintf", "fwscanf", "getwc", "getwchar", "mbrlen",
+    "mbrtowc", "mbsinit", "mbsrtowcs", "putwc", "putwchar", "swprintf", "swscanf", "ungetwc", "vfwprintf",
+    "vfwscanf", "vwprintf", "vwscanf", "wcrtomb", "wcscat", "wcschr", "wcscmp", "wcscoll", "wcscpy", "wcscspn",
+    "wcsftime", "wcslen", "wcsncat", "wcsncmp", "wcsncpy", "wcspbrk", "wcsrchr", "wcsrtombs", "wcsspn", "wcsstr",
+    "wcstod", "wcstok", "wcstol", "wcstombs", "wcstoul", "wcsxfrm", "wctob", "wmemchr", "wmemcmp", "wmemcpy",
+    "wmemmove", "wmemset", "wprintf", "wscanf",
     # WCTYPE
-    'iswalnum', 'iswalpha', 'iswcntrl', 'iswdigit', 'iswgraph', 'iswlower', 'iswprint', 'iswpunct', 'iswspace',
-    'iswupper', 'iswxdigit', 'towlower', 'towupper', 'iswctype', 'wctype',
+    "iswalnum", "iswalpha", "iswcntrl", "iswdigit", "iswgraph", "iswlower", "iswprint", "iswpunct", "iswspace",
+    "iswupper", "iswxdigit", "towlower", "towupper", "iswctype", "wctype",
     # STDDEF
-    'NULL', 'offsetof', 'ptrdiff_t', 'size_t', 'wchar_t',
+    "NULL", "offsetof", "ptrdiff_t", "size_t", "wchar_t",
     # STDARG
-    'va_arg', 'va_end', 'va_start',
+    "va_arg", "va_end", "va_start",
     # DLFCN
-    'dlclose', 'dlerror', 'dlopen', 'dlsym',
+    "dlclose", "dlerror", "dlopen", "dlsym",
     # DIRENT
-    'closedir', 'opendir', 'readdir',
+    "closedir", "opendir", "readdir",
     # SYS_SENDFILE
-    'sendfile',
+    "sendfile",
     # SYS_MMAN
-    'mmap', 'mprotect', 'munmap',
+    "mmap", "mprotect", "munmap",
     # SYS_UTSNAME
-    'uname',
+    "uname",
     # LINK
-    'dladdr'
+    "dladdr"
 }
 
+# https://cs.android.com/search?q=oemcrypto&sq=&ss=android%2Fplatform%2Fsuperproject
 OEM_CRYPTO_API = {
     # Mapping of function names across different API levels (obfuscated names may vary).
-    'rnmsglvj', 'polorucp', 'kqzqahjq', 'pldrclfq', 'kgaitijd', 'cwkfcplc', 'crhqcdet', 'ulns', 'dnvffnze', 'ygjiljer',
-    'qbjxtubz', 'qkfrcjtw', 'rbhjspoh', 'zgtjmxko', 'igrqajte', 'ofskesua', 'qllcoacg', 'pukctkiv', 'ehdqmfmd',
-    'xftzvkwx', 'gndskkuk', 'wcggmnnx', 'kaatohcz', 'ktmgdchz', 'jkcwonus', 'ehmduqyt', 'vewtuecx', 'mxrbzntq',
-    'isyowgmp', 'flzfkhbc'
+    "rnmsglvj", "polorucp", "kqzqahjq", "pldrclfq", "kgaitijd", "cwkfcplc", "crhqcdet", "ulns", "dnvffnze", "ygjiljer",
+    "qbjxtubz", "qkfrcjtw", "rbhjspoh", "zgtjmxko", "igrqajte", "ofskesua", "qllcoacg", "pukctkiv", "ehdqmfmd",
+    "xftzvkwx", "gndskkuk", "wcggmnnx", "kaatohcz", "ktmgdchz", "jkcwonus", "ehmduqyt", "vewtuecx", "mxrbzntq",
+    "isyowgmp", "flzfkhbc"
     # Add more as needed for different versions.
 }
 
 # https://developer.android.com/tools/releases/platforms
 CDM_VENDOR_API = {
-    'mediaserver': [
-        Vendor(22, 11, '1.0', r'libwvdrmengine(?:@\S+)?\.so')
+    "mediaserver": [
+        Vendor(22, 11, "1.0", r"libwvdrmengine(?:@\S+)?\.so")
     ],
-    'mediadrmserver': [
-        Vendor(24, 11, '1.0', r'libwvdrmengine(?:@\S+)?\.so')
+    "mediadrmserver": [
+        Vendor(24, 11, "1.0", r"libwvdrmengine(?:@\S+)?\.so")
     ],
-    'android.hardware.drm@1.0-service.widevine': [
-        Vendor(26, 13, '5.1.0', r'libwvhidl(?:@\S+)?\.so')
+    "android.hardware.drm@1.0-service.widevine": [
+        Vendor(26, 13, "5.1.0", r"libwvhidl(?:@\S+)?\.so")
     ],
-    'android.hardware.drm@1.1-service.widevine': [
-        Vendor(28, 14, '14.0.0', r'libwvhidl(?:@\S+)?\.so')
+    "android.hardware.drm@1.1-service.widevine": [
+        Vendor(28, 14, "14.0.0", r"libwvhidl(?:@\S+)?\.so")
     ],
-    'android.hardware.drm@1.2-service.widevine': [
-        Vendor(29, 15, '15.0.0', r'libwvhidl(?:@\S+)?\.so')
+    "android.hardware.drm@1.2-service.widevine": [
+        Vendor(29, 15, "15.0.0", r"libwvhidl(?:@\S+)?\.so")
     ],
-    'android.hardware.drm@1.3-service.widevine': [
-        Vendor(30, 16, '16.0.0', r'libwvhidl(?:@\S+)?\.so')
+    "android.hardware.drm@1.3-service.widevine": [
+        Vendor(30, 16, "16.0.0", r"libwvhidl(?:@\S+)?\.so")
     ],
-    'android.hardware.drm@1.4-service.widevine': [
-        Vendor(31, 16, '16.1.0', r'libwvhidl(?:@\S+)?\.so')
+    "android.hardware.drm@1.4-service.widevine": [
+        Vendor(31, 16, "16.1.0", r"libwvhidl(?:@\S+)?\.so")
     ],
-    'android.hardware.drm-service.widevine': [
-        Vendor(33, 17, '17.0.0', r'libwvaidl(?:@\S+)?\.so'),
-        Vendor(34, 18, '18.0.0', r'android\.hardware\.drm-service(?:-lazy)?\.widevine(?:@\S+)?'),
-        Vendor(35, 18, '19.0.1', r'android\.hardware\.drm-service(?:-lazy)?\.widevine(?:@\S+)?')
+    "android.hardware.drm-service.widevine": [
+        Vendor(33, 17, "17.0.0", r"libwvaidl(?:@\S+)?\.so"),
+        Vendor(34, 18, "18.0.0", r"android\.hardware\.drm-service(?:-lazy)?\.widevine(?:@\S+)?"),
+        Vendor(35, 18, "19.0.1", r"android\.hardware\.drm-service(?:-lazy)?\.widevine(?:@\S+)?")
     ]
 }
 
 # https://developers.google.com/widevine
 CDM_FUNCTION_API = {
-    'UsePrivacyMode',
-    'GetCdmClientPropertySet',
-    'PrepareKeyRequest',
-    'getOemcryptoDeviceId',
-    'lcc07',
-    'oecc07',
-    'Read',
-    'x1c36',
-    'runningcrc'
+    "UsePrivacyMode",
+    "GetCdmClientPropertySet",
+    "PrepareKeyRequest",
+    "getOemcryptoDeviceId",
+    "lcc07",
+    "oecc07",
+    "Read",
+    "x1c36",
+    "runningcrc"
 }
 
+# Maximum clear API level for Keybox
+# https://cs.android.com/android/platform/superproject/+/android14-qpr3-release:trusty/user/app/sample/hwcrypto/keybox/keybox.c
 KEYBOX_MAX_CLEAR_API = 28
 
 # https://github.com/kaltura/kaltura-device-info-android
 DRM_PLAYER = {
-    'package': 'com.kaltura.kalturadeviceinfo',
-    'path': Path(__file__).parent.parent / 'docs' / 'server' / 'kaltura.apk',
-    'url': 'https://github.com/kaltura/kaltura-device-info-android/releases/download/t3/kaltura-device-info-release.apk'
+    "package": "com.kaltura.kalturadeviceinfo",
+    "path": Path(__file__).parent.parent / "docs" / "server" / "kaltura.apk",
+    "url": "https://github.com/kaltura/kaltura-device-info-android/releases/download/t3/kaltura-device-info-release.apk"
 }